Transform kaba
